Move tests folder outside src and rename to test
